check for identity layout; split pass to separate source
